+++ README.md
@@ -26,27 +26,26 @@ For the full set of developer tools, check out the brand new
 // 1. Initalize market with parameters and metadata
 const market = await SolendMarket.initialize(
   connection,
-  'production', // optional environment argument
-  new PublicKey('7RCz8wb6WXxUhAigok9ttgrVgDFFFbibcirECzWSBauM'), // optional market address (TURBO SOL). Defaults to 'Main' market
+  "production", // optional environment argument
+  new PublicKey("7RCz8wb6WXxUhAigok9ttgrVgDFFFbibcirECzWSBauM") // optional market address (TURBO SOL). Defaults to 'Main' market
 );
-console.log(market.reserves.map(reserve => reserve.config.loanToValueRatio));
+console.log(market.reserves.map((reserve) => reserve.config.loanToValueRatio));
 
 // 2. Read on-chain accounts for reserve data and cache
 await market.loadReserves();
 
-const usdcReserve = market.reserves.find(res => res.config.symbol === 'USDC');
+const usdcReserve = market.reserves.find((res) => res.config.symbol === "USDC");
 console.log(usdcReserve.stats.totalDepositsWads.toString());
 
 // Read Solend liquidity mining stats
-await market.loadRewards()
+await market.loadRewards();
 console.log(reserve.stats.totalSupplyAPY().rewards); // {apy: 0.07, rewardMint: "SLND...
 
 // Refresh all cached data
-market.refreshAll()
+market.refreshAll();
 
-const obligation = market.fetchObligationByWallet('[WALLET_ID]');
-console.log(obligation.stats.borrowLimit)
-
+const obligation = market.fetchObligationByWallet("[WALLET_ID]");
+console.log(obligation.stats.borrowLimit);
 ```
 
 ### Perform lending action
@@ -59,7 +58,7 @@ const solendAction = await SolendAction.buildDepositTxns(
   symbol,
   publicKey,
   "production",
-  new PublicKey('7RCz8wb6WXxUhAigok9ttgrVgDFFFbibcirECzWSBauM'), // optional market address (TURBO SOL). Defaults to 'Main' market
+  new PublicKey("7RCz8wb6WXxUhAigok9ttgrVgDFFFbibcirECzWSBauM") // optional market address (TURBO SOL). Defaults to 'Main' market
 );
 
 await solendAction.sendTransactions(sendTransaction); // sendTransaction from wallet adapter or custom
