+++ src/classes/constants.ts
