+++ src/classes/index.ts
@@ -1,3 +1,5 @@
 export * from "./action";
 export * from "./market";
+export * from "./reserve";
+export * from "./obligation";
 export * from "./types";
