+++ src/classes/market.ts
@@ -1,11 +1,9 @@
-/* eslint-disable max-classes-per-file */
-import { AccountInfo, Connection, PublicKey } from "@solana/web3.js";
+import { Connection, PublicKey } from "@solana/web3.js";
 import { ConfigType } from "./types";
 import BigNumber from "bignumber.js";
-import { parseReserve } from "../state/reserve";
-import { Obligation, parseObligation } from "../state/obligation";
-import BN from "bn.js";
-import { WAD, WANG, SLOTS_PER_YEAR } from "./constants";
+import { SolendObligation } from "./obligation";
+import { SolendReserve } from "./reserve";
+import { parseObligation } from "../state/obligation";
 import axios from "axios";
 
 export type RewardInfo = {
@@ -22,7 +20,7 @@ export type RewardsData = {
   };
 };
 
-type RewardStatType = {
+export type RewardStatType = {
   rewardsPerShare: string;
   totalBalance: string;
   lastSlot: number;
@@ -36,22 +34,23 @@ type RewardStatType = {
 type ExternalRewardStatType = RewardStatType & {
   rewardMint: string;
   rewardSymbol: string;
+  reserveID: string;
+  side: "supply" | "borrow";
 };
-type RewardResponse = {
+
+export type RewardResponse = {
   supply: RewardStatType;
   borrow: RewardStatType;
 };
 
-type ExternalRewardResponse = {
-  supply: ExternalRewardStatType;
-  borrow: ExternalRewardStatType;
-};
-
-type FormattedMarketConfig = ReturnType<typeof formatReserveConfig>;
+export type FormattedMarketConfig = ReturnType<typeof formatReserveConfig>;
 
 const API_ENDPOINT = "https://api.solend.fi";
 
-function formatReserveConfig(config: ConfigType, marketAddress?: string) {
+export function formatReserveConfig(
+  config: ConfigType,
+  marketAddress?: string
+) {
   const market = marketAddress
     ? config.markets.find((mar) => mar.address === marketAddress)
     : config.markets.find((mar) => mar.isPrimary) ?? config.markets[0];
@@ -114,7 +113,7 @@ export class SolendMarket {
     ).data) as ConfigType;
     market.config = formatReserveConfig(rawConfig, marketAddress);
     market.reserves = market.config.reserves.map(
-      (res) => new SolendReserve(res, market, connection)
+      (res) => new SolendReserve(res, connection)
     );
 
     return market;
@@ -123,9 +122,7 @@ export class SolendMarket {
   async fetchObligationByWallet(publicKey: PublicKey) {
     const { config, reserves } = this;
     if (!config) {
-      throw Error(
-        "Market must be initialized to call fetchObligationByWallet."
-      );
+      throw Error("Market must be initialized to call initialize.");
     }
     const obligationAddress = await PublicKey.createWithSeed(
       publicKey,
@@ -171,9 +168,11 @@ export class SolendMarket {
 
   private async loadLMRewardData() {
     const data = (
-      await axios.get(`${API_ENDPOINT}/liquidity-mining/reward-stats`)
+      await axios.get(`${API_ENDPOINT}/liquidity-mining/reward-stats-v2`)
     ).data as Promise<{
-      [key: string]: RewardResponse;
+      [marketAddress: string]: {
+        [mintAddress: string]: RewardResponse;
+      };
     }>;
 
     return data;
@@ -181,10 +180,10 @@ export class SolendMarket {
 
   private async loadExternalRewardData() {
     const data = (
-      await axios.get(`${API_ENDPOINT}/liquidity-mining/external-reward-stats`)
-    ).data as Promise<{
-      [key: string]: ExternalRewardResponse;
-    }>;
+      await axios.get(
+        `${API_ENDPOINT}/liquidity-mining/external-reward-stats-v2?flat=true`
+      )
+    ).data as Promise<Array<ExternalRewardStatType>>;
 
     return data;
   }
@@ -226,6 +225,10 @@ export class SolendMarket {
   }
 
   async loadRewards() {
+    if (!this.config) {
+      throw Error("Market must be initialized to call initialize.");
+    }
+
     const promises = [
       this.loadLMRewardData(),
       this.loadExternalRewardData(),
@@ -236,12 +239,9 @@ export class SolendMarket {
       promises
     );
 
-    const querySymbols = Object.values(externalRewards)
-      .flatMap((reward) => [
-        reward.supply?.rewardSymbol,
-        reward.borrow?.rewardSymbol,
-      ])
-      .filter((x) => x);
+    const querySymbols = [
+      ...new Set(externalRewards.map((reward) => reward.rewardSymbol)),
+    ];
 
     const priceData = await this.loadPriceData(querySymbols.concat("SLND"));
 
@@ -249,8 +249,7 @@ export class SolendMarket {
       const {
         config: { mintAddress },
       } = reserve;
-      const lmReward = lmRewards[mintAddress];
-      const externalReward = externalRewards[mintAddress];
+      const lmReward = lmRewards[this.config!.address][mintAddress];
 
       const supply = [
         lmReward?.supply
@@ -264,18 +263,22 @@ export class SolendMarket {
               price: new BigNumber(priceData.SLND).toNumber(),
             }
           : null,
-        externalReward?.supply
-          ? {
-              rewardRate: this.getLatestRewardRate(
-                externalReward.supply.rewardRates,
-                currentSlot
-              ).rewardRate,
-              rewardMint: externalReward.supply.rewardMint,
-              rewardSymbol: externalReward.supply.rewardSymbol,
-              price: priceData[externalReward.supply.rewardSymbol],
-            }
-          : null,
-      ].filter((x) => x);
+        ...externalRewards
+          .filter(
+            (externalReward) =>
+              externalReward.reserveID === reserve.config.address &&
+              externalReward.side === "supply"
+          )
+          .map((externalReward) => ({
+            rewardRate: this.getLatestRewardRate(
+              externalReward.rewardRates,
+              currentSlot
+            ).rewardRate,
+            rewardMint: externalReward.rewardMint,
+            rewardSymbol: externalReward.rewardSymbol,
+            price: priceData[externalReward.rewardSymbol],
+          })),
+      ].filter(Boolean);
 
       const borrow = [
         lmReward?.borrow
@@ -289,18 +292,22 @@ export class SolendMarket {
               price: new BigNumber(priceData.SLND).toNumber(),
             }
           : null,
-        externalReward?.borrow
-          ? {
-              rewardRate: this.getLatestRewardRate(
-                externalReward.borrow.rewardRates,
-                currentSlot
-              ).rewardRate,
-              rewardMint: externalReward.borrow.rewardMint,
-              rewardSymbol: externalReward.borrow.rewardSymbol,
-              price: priceData[externalReward.borrow.rewardSymbol],
-            }
-          : null,
-      ].filter((x) => x);
+        ...externalRewards
+          .filter(
+            (externalReward) =>
+              externalReward.reserveID === reserve.config.address &&
+              externalReward.side === "borrow"
+          )
+          .map((externalReward) => ({
+            rewardRate: this.getLatestRewardRate(
+              externalReward.rewardRates,
+              currentSlot
+            ).rewardRate,
+            rewardMint: externalReward.rewardMint,
+            rewardSymbol: externalReward.rewardSymbol,
+            price: priceData[externalReward.rewardSymbol],
+          })),
+      ].filter(Boolean);
 
       return {
         ...acc,
@@ -312,7 +319,7 @@ export class SolendMarket {
     }, {});
 
     const refreshReserves = this.reserves.map((reserve) => {
-      return reserve.load();
+      return reserve.load(this.rewardsData ?? undefined);
     });
 
     await Promise.all(refreshReserves);
@@ -346,431 +353,3 @@ export class SolendMarket {
     await Promise.all(promises);
   }
 }
-
-export type ReserveData = {
-  optimalUtilizationRate: number;
-  loanToValueRatio: number;
-  liquidationBonus: number;
-  liquidationThreshold: number;
-  minBorrowRate: number;
-  optimalBorrowRate: number;
-  maxBorrowRate: number;
-  borrowFeePercentage: number;
-  hostFeePercentage: number;
-  depositLimit: BN;
-  reserveBorrowLimit: BN;
-  name: string;
-  symbol: string;
-  decimals: number;
-  mintAddress: string;
-  totalDepositsWads: BN;
-  totalBorrowsWads: BN;
-  totalLiquidityWads: BN;
-  supplyInterestAPY: number;
-  borrowInterestAPY: number;
-  assetPriceUSD: number;
-  userDepositLimit?: number;
-  cumulativeBorrowRateWads: BN;
-  cTokenExchangeRate: number;
-};
-
-type ParsedReserve = NonNullable<ReturnType<typeof parseReserve>>["info"];
-type FormattedReserveConfig = FormattedMarketConfig["reserves"][0];
-
-export class SolendReserve {
-  config: FormattedReserveConfig;
-
-  private market: SolendMarket;
-
-  private buffer: AccountInfo<Buffer> | null;
-
-  stats: ReserveData | null;
-
-  private connection: Connection;
-
-  constructor(
-    reserveConfig: FormattedReserveConfig,
-    market: SolendMarket,
-    connection: Connection
-  ) {
-    this.config = reserveConfig;
-    this.market = market;
-    this.buffer = null;
-    this.stats = null;
-    this.connection = connection;
-  }
-
-  private calculateSupplyAPY = (reserve: ParsedReserve) => {
-    const apr = this.calculateSupplyAPR(reserve);
-    const apy =
-      new BigNumber(1)
-        .plus(new BigNumber(apr).dividedBy(SLOTS_PER_YEAR))
-        .toNumber() **
-        SLOTS_PER_YEAR -
-      1;
-    return apy;
-  };
-
-  private calculateBorrowAPY = (reserve: ParsedReserve) => {
-    const apr = this.calculateBorrowAPR(reserve);
-    const apy =
-      new BigNumber(1)
-        .plus(new BigNumber(apr).dividedBy(SLOTS_PER_YEAR))
-        .toNumber() **
-        SLOTS_PER_YEAR -
-      1;
-    return apy;
-  };
-
-  private calculateSupplyAPR(reserve: ParsedReserve) {
-    const currentUtilization = this.calculateUtilizationRatio(reserve);
-
-    const borrowAPY = this.calculateBorrowAPR(reserve);
-    return currentUtilization * borrowAPY;
-  }
-
-  private calculateUtilizationRatio(reserve: ParsedReserve) {
-    const totalBorrowsWads = new BigNumber(
-      reserve.liquidity.borrowedAmountWads.toString()
-    ).div(WAD);
-    const currentUtilization = totalBorrowsWads
-      .dividedBy(
-        totalBorrowsWads.plus(reserve.liquidity.availableAmount.toString())
-      )
-      .toNumber();
-
-    return currentUtilization;
-  }
-
-  private calculateBorrowAPR(reserve: ParsedReserve) {
-    const currentUtilization = this.calculateUtilizationRatio(reserve);
-    const optimalUtilization = reserve.config.optimalUtilizationRate / 100;
-
-    let borrowAPR;
-    if (optimalUtilization === 1.0 || currentUtilization < optimalUtilization) {
-      const normalizedFactor = currentUtilization / optimalUtilization;
-      const optimalBorrowRate = reserve.config.optimalBorrowRate / 100;
-      const minBorrowRate = reserve.config.minBorrowRate / 100;
-      borrowAPR =
-        normalizedFactor * (optimalBorrowRate - minBorrowRate) + minBorrowRate;
-    } else {
-      const normalizedFactor =
-        (currentUtilization - optimalUtilization) / (1 - optimalUtilization);
-      const optimalBorrowRate = reserve.config.optimalBorrowRate / 100;
-      const maxBorrowRate = reserve.config.maxBorrowRate / 100;
-      borrowAPR =
-        normalizedFactor * (maxBorrowRate - optimalBorrowRate) +
-        optimalBorrowRate;
-    }
-
-    return borrowAPR;
-  }
-
-  setBuffer(buffer: AccountInfo<Buffer> | null) {
-    this.buffer = buffer;
-  }
-
-  async load() {
-    if (!this.buffer) {
-      this.buffer = await this.connection.getAccountInfo(
-        new PublicKey(this.config.address),
-        "processed"
-      );
-    }
-
-    if (!this.buffer) {
-      throw Error(`Error requesting account info for ${this.config.name}`);
-    }
-
-    const parsedData = parseReserve(
-      new PublicKey(this.config.address),
-      this.buffer
-    )?.info;
-    if (!parsedData) {
-      throw Error(`Unable to parse data of reserve ${this.config.name}`);
-    }
-
-    this.stats = await this.formatReserveData(parsedData);
-  }
-
-  calculateRewardAPY(
-    rewardRate: string,
-    poolSize: string,
-    rewardPrice: number,
-    tokenPrice: number,
-    decimals: number
-  ) {
-    const poolValueUSD = new BigNumber(poolSize)
-      .times(tokenPrice)
-      .dividedBy("1".concat(Array(decimals + 1).join("0")))
-      .dividedBy(WAD);
-
-    return new BigNumber(rewardRate)
-      .multipliedBy(rewardPrice)
-      .dividedBy(poolValueUSD)
-      .dividedBy(WANG);
-  }
-
-  totalSupplyAPY() {
-    const { stats } = this;
-    if (!this.market.rewardsData || !stats) {
-      throw Error(
-        "SolendMarket must be initialized with the withRewardData flag as true and load must be called on the reserve."
-      );
-    }
-
-    const rewards = this.market.config?.isPrimary
-      ? this.market.rewardsData[this.config.mintAddress].supply.map(
-          (reward) => ({
-            rewardMint: reward.rewardMint,
-            rewardSymbol: reward.rewardSymbol,
-            apy: this.calculateRewardAPY(
-              reward.rewardRate,
-              stats.totalDepositsWads.toString(),
-              reward.price,
-              stats.assetPriceUSD,
-              this.config.decimals
-            ).toNumber(),
-            price: reward.price,
-          })
-        )
-      : [];
-
-    const totalAPY = new BigNumber(stats.supplyInterestAPY)
-      .plus(
-        rewards.reduce((acc, reward) => acc.plus(reward.apy), new BigNumber(0))
-      )
-      .toNumber();
-
-    return {
-      interestAPY: stats.supplyInterestAPY,
-      totalAPY,
-      rewards,
-    };
-  }
-
-  totalBorrowAPY() {
-    const { stats } = this;
-    if (!this.market.rewardsData || !stats) {
-      throw Error(
-        "SolendMarket must be initialized with the withRewardData flag as true and load must be called on the reserve."
-      );
-    }
-
-    const rewards = this.market.config?.isPrimary
-      ? this.market.rewardsData[this.config.mintAddress].borrow.map(
-          (reward) => ({
-            rewardMint: reward.rewardMint,
-            rewardSymbol: reward.rewardSymbol,
-            apy: this.calculateRewardAPY(
-              reward.rewardRate,
-              stats.totalBorrowsWads.toString(),
-              reward.price,
-              stats.assetPriceUSD,
-              this.config.decimals
-            ).toNumber(),
-            price: reward.price,
-          })
-        )
-      : [];
-    const totalAPY = new BigNumber(stats.borrowInterestAPY)
-      .minus(
-        rewards.reduce((acc, reward) => acc.plus(reward.apy), new BigNumber(0))
-      )
-      .toNumber();
-
-    return {
-      interestAPY: stats.borrowInterestAPY,
-      totalAPY,
-      rewards,
-    };
-  }
-
-  private formatReserveData(
-    parsedData: NonNullable<ReturnType<typeof parseReserve>>["info"]
-  ): ReserveData {
-    const totalBorrowsWads = parsedData.liquidity.borrowedAmountWads;
-    const totalLiquidityWads = parsedData.liquidity.availableAmount.mul(
-      new BN(WAD)
-    );
-    const totalDepositsWads = totalBorrowsWads.add(totalLiquidityWads);
-    const cTokenExchangeRate = new BigNumber(totalDepositsWads.toString())
-      .div(parsedData.collateral.mintTotalSupply.toString())
-      .div(WAD)
-      .toNumber();
-
-    return {
-      // Reserve config
-      optimalUtilizationRate: parsedData.config.optimalUtilizationRate / 100,
-      loanToValueRatio: parsedData.config.loanToValueRatio / 100,
-      liquidationBonus: parsedData.config.liquidationBonus / 100,
-      liquidationThreshold: parsedData.config.liquidationThreshold / 100,
-      minBorrowRate: parsedData.config.minBorrowRate / 100,
-      optimalBorrowRate: parsedData.config.optimalBorrowRate / 100,
-      maxBorrowRate: parsedData.config.maxBorrowRate / 100,
-      borrowFeePercentage: new BigNumber(
-        parsedData.config.fees.borrowFeeWad.toString()
-      )
-        .dividedBy(WAD)
-        .toNumber(),
-      hostFeePercentage: parsedData.config.fees.hostFeePercentage / 100,
-      depositLimit: parsedData.config.depositLimit,
-      reserveBorrowLimit: parsedData.config.borrowLimit,
-
-      // Reserve info
-      name: this.config.name,
-      symbol: this.config.symbol,
-      decimals: this.config.decimals,
-      mintAddress: this.config.mintAddress,
-      totalDepositsWads,
-      totalBorrowsWads,
-      totalLiquidityWads,
-      supplyInterestAPY: this.calculateSupplyAPY(parsedData),
-      borrowInterestAPY: this.calculateBorrowAPY(parsedData),
-      assetPriceUSD: new BigNumber(parsedData.liquidity.marketPrice.toString())
-        .div(WAD)
-        .toNumber(),
-      userDepositLimit: this.config.userSupplyCap,
-      cumulativeBorrowRateWads: parsedData.liquidity.cumulativeBorrowRateWads,
-      cTokenExchangeRate,
-    };
-  }
-}
-
-export type Position = {
-  mintAddress: string;
-  amount: BN;
-};
-
-export type ObligationStats = {
-  liquidationThreshold: number;
-  userTotalDeposit: number;
-  userTotalBorrow: number;
-  borrowLimit: number;
-  borrowUtilization: number;
-  netAccountValue: number;
-  positions: number;
-};
-
-export class SolendObligation {
-  walletAddress: PublicKey;
-
-  obligationAddress: PublicKey;
-
-  deposits: Array<Position>;
-
-  borrows: Array<Position>;
-
-  obligationStats: ObligationStats;
-
-  constructor(
-    walletAddress: PublicKey,
-    obligationAddress: PublicKey,
-    obligation: Obligation,
-    reserves: Array<SolendReserve>
-  ) {
-    this.walletAddress = walletAddress;
-    this.obligationAddress = obligationAddress;
-
-    const positionDetails = this.calculatePositions(obligation, reserves);
-
-    this.deposits = positionDetails.deposits;
-    this.borrows = positionDetails.borrows;
-    this.obligationStats = positionDetails.stats;
-  }
-
-  private calculatePositions(
-    obligation: Obligation,
-    reserves: Array<SolendReserve>
-  ) {
-    let userTotalDeposit = new BigNumber(0);
-    let borrowLimit = new BigNumber(0);
-    let liquidationThreshold = new BigNumber(0);
-    let positions = 0;
-
-    const deposits = obligation.deposits.map((deposit) => {
-      const reserve = reserves.find(
-        (reserve) =>
-          reserve.config.address === deposit.depositReserve.toBase58()
-      );
-      const loanToValue = reserve!.stats!.loanToValueRatio;
-      const liqThreshold = reserve!.stats!.liquidationThreshold;
-
-      const supplyAmount = new BN(
-        Math.floor(
-          new BigNumber(deposit.depositedAmount.toString())
-            .multipliedBy(reserve!.stats!.cTokenExchangeRate)
-            .toNumber()
-        )
-      );
-      const supplyAmountUSD = new BigNumber(supplyAmount.toString())
-        .multipliedBy(reserve!.stats!.assetPriceUSD)
-        .dividedBy("1".concat(Array(reserve!.stats!.decimals + 1).join("0")));
-
-      userTotalDeposit = userTotalDeposit.plus(supplyAmountUSD);
-
-      borrowLimit = borrowLimit.plus(supplyAmountUSD.multipliedBy(loanToValue));
-
-      liquidationThreshold = liquidationThreshold.plus(
-        supplyAmountUSD.multipliedBy(liqThreshold)
-      );
-
-      if (!supplyAmount.eq(new BN("0"))) {
-        positions += 1;
-      }
-
-      return {
-        mintAddress: reserve!.config.mintAddress,
-        amount: supplyAmount,
-      };
-    });
-
-    let userTotalBorrow = new BigNumber(0);
-
-    const borrows = obligation.borrows.map((borrow) => {
-      const reserve = reserves.find(
-        (reserve) => reserve.config.address === borrow.borrowReserve.toBase58()
-      );
-
-      const borrowAmount = new BN(
-        Math.floor(
-          new BigNumber(borrow.borrowedAmountWads.toString())
-            .multipliedBy(reserve!.stats!.cumulativeBorrowRateWads.toString())
-            .dividedBy(borrow.cumulativeBorrowRateWads.toString())
-            .dividedBy(WAD)
-            .toNumber()
-        ).toString()
-      );
-
-      const borrowAmountUSD = new BigNumber(borrowAmount.toString())
-        .multipliedBy(reserve!.stats!.assetPriceUSD)
-        .dividedBy("1".concat(Array(reserve!.stats!.decimals + 1).join("0")));
-
-      if (!borrowAmount.eq(new BN("0"))) {
-        positions += 1;
-      }
-
-      userTotalBorrow = userTotalBorrow.plus(borrowAmountUSD);
-
-      return {
-        mintAddress: reserve!.config.mintAddress,
-        amount: borrowAmount,
-      };
-    });
-
-    return {
-      deposits,
-      borrows,
-      stats: {
-        liquidationThreshold: liquidationThreshold.toNumber(),
-        userTotalDeposit: userTotalDeposit.toNumber(),
-        userTotalBorrow: userTotalBorrow.toNumber(),
-        borrowLimit: borrowLimit.toNumber(),
-        borrowUtilization: userTotalBorrow.dividedBy(borrowLimit).toNumber(),
-        netAccountValue: userTotalDeposit.minus(userTotalBorrow).toNumber(),
-        positions,
-      },
-    };
-  }
-}
-/* eslint-enable max-classes-per-file */
